requirements.txt
# requirements.txt
praw>=7.7.1
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.66.1